@@ -14,30 +14,30 @@ dependencies:
   cupertino_icons: ^1.0.2
   dartz: ^0.10.1
   dio: ^5.3.0
-  flutter_bloc: ^8.1.3
+  flutter_bloc: ^9.1.1
   equatable: ^2.0.5
-  get_it: ^7.6.0
+  get_it: ^8.0.3
   hive: ^2.2.3
   hive_flutter: ^1.1.0
   path_provider: ^2.0.11
   flutter_svg: ^2.0.5
-  camera: ^0.10.0
-  image_picker: ^0.8.6+1
-  tflite_flutter: ^0.10.0
+  camera: ^0.11.2
+  image_picker: ^1.1.2
+  tflite_flutter: ^0.11.0
   image: ^4.0.15
-  logger: ^1.3.0
+  logger: ^2.6.1
   provider: ^6.0.5
   shared_preferences: ^2.1.0
-  google_fonts: ^4.0.4
+  google_fonts: ^6.2.1
   sqflite: ^2.3.0
-  connectivity_plus: ^5.0.2
+  connectivity_plus: ^6.1.4
   path: ^1.8.3
-  bloc: ^8.1.4
+  bloc: ^9.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^6.0.0
   hive_generator: ^2.0.0
   build_runner: ^2.3.3
 
