@@ -5,7 +5,7 @@ import 'package:lit_eracy/presentation/blocs/content/content_bloc.dart';
 import 'package:lit_eracy/presentation/widgets/lesson_card.dart';
 
 class HomePage extends StatelessWidget {
-  const HomePage({Key? key}) : super(key: key);
+  const HomePage({super.key});
 
   @override
   Widget build(BuildContext context) {
