@@ -1,5 +1,5 @@
 class AppConstants {
-  static const String apiBaseUrl = 'http://localhost:8000';
+  static const String apiBaseUrl = 'http://localhost:8000/api';
   static const int splashDelaySeconds = 2;
   static const String dbName = 'ai_literacy_app.db';
   static const String hiveBoxUser = 'userBox';
